Give offsettable its own file
