+++ src/griddyn/gridComponent.cpp
@@ -24,153 +24,7 @@
 
 namespace griddyn
 {
-// this functions is here since it depends on gridComponent information
-Lp offsetTable::getLocations (const stateData &sD,
-                              double dest[],
-                              const solverMode &sMode,
-                              const gridComponent *comp) const
-{
-    Lp Loc = getLocations (sD, sMode, comp);
-    if ((sMode.local) || (sD.empty ()))
-    {
-        Loc.destLoc =
-          (dest == nullptr) ? const_cast<double *> (comp->m_state.data ()) + offsetContainer[0].algOffset : dest;
-        Loc.destDiffLoc = Loc.destLoc + Loc.algSize;
-    }
-    else if (isDAE (sMode))
-    {
-        Loc.destLoc = dest + Loc.algOffset;
-        Loc.destDiffLoc = dest + Loc.diffOffset;
-    }
-    else if (hasAlgebraic (sMode))
-    {
-        Loc.destLoc = dest + Loc.algOffset;
-        Loc.destDiffLoc = nullptr;
-    }
-    else if (hasDifferential (sMode))
-    {
-        Loc.destDiffLoc = dest + Loc.diffOffset;
-        Loc.destLoc = nullptr;
-    }
-    else
-    {
-        Loc.destLoc =
-          (dest == nullptr) ? const_cast<double *> (comp->m_state.data ()) + offsetContainer[0].algOffset : dest;
-        Loc.destDiffLoc = Loc.destLoc + Loc.algSize;
-    }
-    return Loc;
-}
 
-// this functions is here since it depends on gridComponent information
-Lp offsetTable::getLocations (const stateData &sD, const solverMode &sMode, const gridComponent *comp) const
-{
-    Lp Loc;
-    Loc.algOffset = offsetContainer[sMode.offsetIndex].algOffset;
-    Loc.diffOffset = offsetContainer[sMode.offsetIndex].diffOffset;
-    Loc.diffSize = offsetContainer[sMode.offsetIndex].total.diffSize;
-    Loc.algSize = offsetContainer[sMode.offsetIndex].total.algSize;
-    if ((sMode.local) || (sD.empty ()))
-    {
-        Loc.time = comp->prevTime;
-        Loc.algStateLoc = comp->m_state.data ();
-        Loc.diffStateLoc = comp->m_state.data () + Loc.algSize;
-        Loc.dstateLoc = comp->m_dstate_dt.data () + Loc.algSize;
-        if (Loc.algOffset == kNullLocation)
-        {
-            Loc.algOffset = 0;
-        }
-        if (Loc.diffOffset == kNullLocation)
-        {
-            Loc.diffOffset = Loc.algSize;
-        }
-    }
-    else if (isDAE (sMode))
-    {
-        Loc.time = sD.time;
-        Loc.algStateLoc = sD.state + Loc.algOffset;
-        Loc.diffStateLoc = sD.state + Loc.diffOffset;
-        Loc.dstateLoc = sD.dstate_dt + Loc.diffOffset;
-    }
-    else if (hasAlgebraic (sMode))
-    {
-        Loc.time = sD.time;
-        if (sD.state != nullptr)
-        {
-            Loc.algStateLoc = sD.state + Loc.algOffset;
-        }
-        else
-        {
-            Loc.algStateLoc = sD.algState + Loc.algOffset;
-        }
-        if ((isDynamic (sMode)) && (sD.pairIndex != kNullLocation))
-        {
-            if (sD.diffState != nullptr)
-            {
-                Loc.diffStateLoc = sD.diffState + offsetContainer[sD.pairIndex].diffOffset;
-            }
-            else if (sD.fullState != nullptr)
-            {
-                Loc.diffStateLoc = sD.fullState + offsetContainer[sD.pairIndex].diffOffset;
-            }
-
-            if (sD.dstate_dt != nullptr)
-            {
-                Loc.dstateLoc = sD.dstate_dt + offsetContainer[sD.pairIndex].diffOffset;
-            }
-        }
-        else
-        {
-            Loc.diffStateLoc = comp->m_state.data () + offsetContainer[0].diffOffset;
-            Loc.dstateLoc = comp->m_dstate_dt.data () + offsetContainer[0].diffOffset;
-        }
-        Loc.destDiffLoc = nullptr;
-    }
-    else if (hasDifferential (sMode))
-    {
-        Loc.time = sD.time;
-        if (sD.state != nullptr)
-        {
-            Loc.diffStateLoc = sD.state + Loc.diffOffset;
-        }
-        else
-        {
-            Loc.diffStateLoc = sD.diffState + Loc.diffOffset;
-        }
-        Loc.dstateLoc = sD.dstate_dt + Loc.diffOffset;
-        if (sD.pairIndex != kNullLocation)
-        {
-            if (sD.algState != nullptr)
-            {
-                Loc.algStateLoc = sD.algState + offsetContainer[sD.pairIndex].algOffset;
-            }
-            else if (sD.fullState != nullptr)
-            {
-                Loc.algStateLoc = sD.fullState + offsetContainer[sD.pairIndex].algOffset;
-            }
-        }
-        else
-        {
-            Loc.algStateLoc = comp->m_state.data () + offsetContainer[0].algOffset;
-        }
-        Loc.destLoc = nullptr;
-    }
-    else
-    {
-        Loc.time = comp->prevTime;
-        Loc.algStateLoc = comp->m_state.data ();
-        Loc.diffStateLoc = comp->m_state.data () + Loc.algSize;
-        Loc.dstateLoc = comp->m_dstate_dt.data () + Loc.algSize;
-        if (Loc.algOffset == kNullLocation)
-        {
-            Loc.algOffset = 0;
-        }
-        if (Loc.diffOffset == kNullLocation)
-        {
-            Loc.diffOffset = Loc.algSize;
-        }
-    }
-    return Loc;
-}
 
 gridComponent::gridComponent (const std::string &objName) : coreObject (objName)
 {
