+++ src/griddyn/gridComponentHelperClasses.cpp
@@ -321,319 +321,4 @@ void solverOffsets::setOffset (index_t newOffset)
     }
 }
 
-
-offsetTable::offsetTable () noexcept : offsetContainer (DEFAULT_OFFSET_CONTAINER_SIZE)
-{
-    // most simulations use the first 1 and powerflow(2) and likely dynamic DAE(3)  and often 4 and 5 for dynamic
-    // partitioned
-#if DEFAULT_OFFSET_CONTAINER_SIZE==0
-    offsetContainer.resize(1);
-#endif
-    offsetContainer[0].sMode = cLocalSolverMode;
-}
-
-bool offsetTable::isLoaded (const solverMode &sMode) const
-{
-    return (isValidIndex (sMode.offsetIndex)) &&
-           ((offsetContainer[sMode.offsetIndex].stateLoaded) && (offsetContainer[sMode.offsetIndex].rootsLoaded) &&
-            (offsetContainer[sMode.offsetIndex].jacobianLoaded));
-}
-
-bool offsetTable::isStateCountLoaded (const solverMode &sMode) const
-{
-    return isValidIndex (sMode.offsetIndex) && offsetContainer[sMode.offsetIndex].stateLoaded;
-}
-
-bool offsetTable::isRootCountLoaded (const solverMode &sMode) const
-{
-    return isValidIndex (sMode.offsetIndex) && offsetContainer[sMode.offsetIndex].rootsLoaded;
-}
-
-bool offsetTable::isJacobianCountLoaded(const solverMode &sMode) const
-{
-    return isValidIndex(sMode.offsetIndex) && offsetContainer[sMode.offsetIndex].jacobianLoaded;
-}
-
-solverOffsets &offsetTable::getOffsets (const solverMode &sMode)
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        offsetContainer.resize (sMode.offsetIndex + 1);
-        offsetContainer[sMode.offsetIndex].sMode = sMode;
-    }
-    return offsetContainer[sMode.offsetIndex];
-}
-
-const solverOffsets &offsetTable::getOffsets (const solverMode &sMode) const
-{
-    return isValidIndex (sMode.offsetIndex) ? offsetContainer[sMode.offsetIndex] : nullOffsets;
-}
-
-void offsetTable::setOffsets (const solverOffsets &newOffsets, const solverMode &sMode)
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        offsetContainer.resize (sMode.offsetIndex + 1);
-    }
-    offsetContainer[sMode.offsetIndex].sMode = sMode;
-    offsetContainer[sMode.offsetIndex].setOffsets (newOffsets);
-}
-
-void offsetTable::setOffset (index_t newOffset, const solverMode &sMode)
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        offsetContainer.resize (sMode.offsetIndex + 1);
-    }
-    offsetContainer[sMode.offsetIndex].sMode = sMode;
-    offsetContainer[sMode.offsetIndex].setOffset (newOffset);
-}
-
-void offsetTable::setAlgOffset (index_t newOffset, const solverMode &sMode)
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        offsetContainer.resize (sMode.offsetIndex + 1);
-    }
-    offsetContainer[sMode.offsetIndex].sMode = sMode;
-    offsetContainer[sMode.offsetIndex].algOffset = newOffset;
-}
-
-void offsetTable::setDiffOffset (index_t newOffset, const solverMode &sMode)
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        offsetContainer.resize (sMode.offsetIndex + 1);
-    }
-    offsetContainer[sMode.offsetIndex].sMode = sMode;
-    offsetContainer[sMode.offsetIndex].diffOffset = newOffset;
-}
-
-void offsetTable::setVOffset (index_t newOffset, const solverMode &sMode)
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        offsetContainer.resize (sMode.offsetIndex + 1);
-    }
-    offsetContainer[sMode.offsetIndex].sMode = sMode;
-    offsetContainer[sMode.offsetIndex].vOffset = newOffset;
-}
-
-void offsetTable::setAOffset (index_t newOffset, const solverMode &sMode)
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        offsetContainer.resize (sMode.offsetIndex + 1);
-    }
-    offsetContainer[sMode.offsetIndex].sMode = sMode;
-    offsetContainer[sMode.offsetIndex].aOffset = newOffset;
-}
-
-void offsetTable::setRootOffset (index_t newOffset, const solverMode &sMode)
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        offsetContainer.resize (sMode.offsetIndex + 1);
-    }
-    offsetContainer[sMode.offsetIndex].sMode = sMode;
-    offsetContainer[sMode.offsetIndex].rootOffset = newOffset;
-}
-
-index_t offsetTable::maxIndex (const solverMode &sMode) const
-{
-    if (!isValidIndex (sMode.offsetIndex))
-    {
-        return 0;
-    }
-    auto so = offsetContainer[sMode.offsetIndex];
-    index_t mx;
-    if (isDynamic (sMode))
-    {
-        mx = so.diffOffset + so.total.diffSize;
-        if (so.algOffset + so.total.algSize > mx)
-        {
-            mx = so.algOffset + so.total.algSize;
-        }
-    }
-    else
-    {
-        mx = so.algOffset + so.total.algSize;
-    }
-    if ((so.vOffset != kNullLocation) && (so.vOffset + so.total.vSize > mx))
-    {
-        mx = so.vOffset + so.total.vSize;
-    }
-    if ((so.aOffset != kNullLocation) && (so.aOffset + so.total.aSize > mx))
-    {
-        mx = so.aOffset + so.total.aSize;
-    }
-    return mx;
-}
-
-void offsetTable::getLocations (const solverMode &sMode, Lp *Loc) const
-{
-    Loc->algOffset = offsetContainer[sMode.offsetIndex].algOffset;
-    Loc->diffOffset = offsetContainer[sMode.offsetIndex].diffOffset;
-}
-
-void offsetTable::unload (bool dynamic_only)
-{
-    if (dynamic_only)
-    {
-        for (auto &so : offsetContainer)
-        {
-            if (isDynamic (so.sMode))
-            {
-                so.stateLoaded = false;
-                so.rootsLoaded = false;
-                so.jacobianLoaded = false;
-                so.diffOffset = kNullLocation;
-                so.algOffset = kNullLocation;
-            }
-        }
-    }
-    else
-    {
-        for (auto &so : offsetContainer)
-        {
-            so.stateLoaded = false;
-            so.rootsLoaded = false;
-            so.jacobianLoaded = false;
-            so.diffOffset = kNullLocation;
-            so.algOffset = kNullLocation;
-        }
-    }
-}
-
-void offsetTable::stateUnload (bool dynamic_only)
-{
-    if (dynamic_only)
-    {
-        for (auto &so : offsetContainer)
-        {
-            if (isDynamic (so.sMode))
-            {
-                so.stateLoaded = false;
-                so.diffOffset = kNullLocation;
-                so.algOffset = kNullLocation;
-            }
-        }
-    }
-    else
-    {
-        for (auto &so : offsetContainer)
-        {
-            so.stateLoaded = false;
-            so.diffOffset = kNullLocation;
-            so.algOffset = kNullLocation;
-        }
-    }
-}
-
-void offsetTable::rootUnload (bool dynamic_only)
-{
-    if (dynamic_only)
-    {
-        for (auto &so : offsetContainer)
-        {
-            if (isDynamic (so.sMode))
-            {
-                so.rootsLoaded = false;
-            }
-        }
-    }
-    else
-    {
-        for (auto &so : offsetContainer)
-        {
-            so.rootsLoaded = false;
-        }
-    }
-}
-void offsetTable::JacobianUnload(bool dynamic_only)
-{
-    if (dynamic_only)
-    {
-        for (auto &so : offsetContainer)
-        {
-            if (isDynamic(so.sMode))
-            {
-                so.jacobianLoaded = false;
-            }
-        }
-    }
-    else
-    {
-        for (auto &so : offsetContainer)
-        {
-            so.jacobianLoaded = false;
-        }
-    }
-}
-
-void offsetTable::localUpdateAll (bool dynamic_only)
-{
-    if (dynamic_only)
-    {
-        for (auto &so : offsetContainer)
-        {
-            if (isDynamic (so.sMode))
-            {
-                auto &lc = local ();
-                so.total.algRoots = so.local.algRoots = lc.local.algRoots;
-                so.total.diffRoots = so.local.diffRoots = lc.local.diffRoots;
-                so.total.jacSize = so.local.jacSize = lc.local.jacSize;
-                so.rootsLoaded = true;
-                so.jacobianLoaded = true;
-            }
-        }
-    }
-    else
-    {
-        for (auto &so : offsetContainer)
-        {
-            so.local = local ().local;
-            so.localLoadAll(true);
-        }
-    }
-}
-const solverMode &offsetTable::getSolverMode (index_t index) const
-{
-    return isValidIndex (index) ? offsetContainer[index].sMode : cEmptySolverMode;
-}
-
-const solverMode &offsetTable::find (const solverMode &tMode) const
-{
-    for (auto &so : offsetContainer)
-    {
-        if (so.sMode.dynamic != tMode.dynamic)
-        {
-            continue;
-        }
-        if (so.sMode.local != tMode.local)
-        {
-            continue;
-        }
-        if (so.sMode.algebraic != tMode.algebraic)
-        {
-            continue;
-        }
-        if (so.sMode.differential != tMode.differential)
-        {
-            continue;
-        }
-
-        if (so.sMode.extended_state != tMode.extended_state)
-        {
-            continue;
-        }
-        if (so.sMode.approx != tMode.approx)
-        {
-            continue;
-        }
-        return so.sMode;
-    }
-    return cEmptySolverMode;
-}
-
 }  // namespace griddyn
